Python/requirements copy.txt
python-dotenv
langchain
langchain_core
langchain_community
langchain-ollama
langchain_text_splitters
faiss-cpu
langchain-ollama
beautifulsoup4
html2text
opencv-python
ultralytics
torch
torchvision
numpy
streamlit
uvicorn
fastapi
fitz
pymupdf